+++ nmt-baitap/009_xoa_so.cpp
@@ -6,7 +6,7 @@ int main() {
   long long N;
   cin >> N;
 
-  cout << (long long)N / 1000 << endl;
+  cout << N / 1000 << endl;
 
   return 0;
 }
